# improvement: ReasonablePosixTimeZone#to_offset returns default value directly when dst is none